requirements.txt
PySide6>=6.5.0
pyvistaqt>=0.11.0
pyvista>=0.43.0
vtk>=9.2.0
trimesh>=4.0.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0

# Packaging (only needed for building the .exe, not for running from source)
pyinstaller>=6.0.0
